+++ Team41/Code41/src/spa/src/DesignExtractor/CFGExtractor.cpp
@@ -36,9 +36,9 @@ void CFGExtractor::buildInitCFG() {
             if (parentCFGNode) // IF or WHILE CFGNode point to first stmt in container
                 addCFGEdge(parentCFGNode, childCFGNode); // add forward CFG edge
 
-            for (int i = 0; i < ch.size(); ++i) {
+            for (int i = 0; i < (int) ch.size(); ++i) {
                 TNodeType childType = ch[i]->getType();
-                CFGNode *neighbourCFGNode = (i != ch.size() - 1) ? createCFGNode(ch[i+1])
+                CFGNode *neighbourCFGNode = (i != (int) ch.size() - 1) ? createCFGNode(ch[i+1])
                                                                  : nullptr; // last child has no neighbour
                 if (childType == TNodeType::ifStmt) {
                     bfsQ.push({ch[i], childCFGNode, nullptr});
@@ -79,7 +79,7 @@ void CFGExtractor::linkBackNode() {
         TNodeType type = curTNode->getType();
         if (type == TNodeType::stmtLst) {
             vector<TNode *> ch = curTNode->getChildren();
-            for (int i = 0; i < ch.size() - 1; ++i) {
+            for (int i = 0; i < (int) ch.size() - 1; ++i) {
                 TNodeType childType = ch[i]->getType();
                 if (childType == TNodeType::ifStmt)
                     bfsQ.push({ch[i], ch[i+1]}); // end of IF will link to IF's neighbour node
