+++ Team41/Code41/src/spa/src/QPS/QueryLexer.cpp
@@ -7,13 +7,13 @@ using namespace std;
 QueryLexer::QueryLexer(const string &input) : input(input) {}
 
 int QueryLexer::skipSpaces() {
-    if (index >= input.length()) {
+    if (index >= (int) input.length()) {
         return index;
     }
     char curr = input.at(index);
     while (curr == ' ' || curr == '\n') {
         index++;
-        if (index < input.length()) {
+        if (index < (int) input.length()) {
             curr = input.at(index);
         } else { break; }
     }
@@ -25,7 +25,7 @@ bool QueryLexer::isSpecialChar(char w) {
 }
 
 string QueryLexer::nextToken() {
-    if (index >= input.length()) {
+    if (index >= (int) input.length()) {
         return "";
     }
     string out = "";
@@ -44,7 +44,7 @@ string QueryLexer::nextToken() {
             out.push_back(curr);
         }
         index++;
-        if (index >= input.length()) {
+        if (index >= (int) input.length()) {
             break;
         }
         curr = input.at(index);
@@ -57,7 +57,7 @@ bool QueryLexer::isValidSynonym(string w) {
      IDENT : LETTER ( LETTER | DIGIT )*
      synonym : IDENT
      */
-    for (int i = 0; i < w.length(); i++) {
+    for (int i = 0; i < (int) w.length(); i++) {
         char c = w.at(i);
         if (i == 0) { // first char must be LETTER
             if (!isalpha(c)) {
@@ -235,14 +235,14 @@ optional<string> QueryLexer::nextAttribute() {
 
 bool QueryLexer::isEndOfQuery() {
     skipSpaces();
-    return index == input.length();
+    return index == (int) input.length();
 }
 
 bool QueryLexer::peekNextIsString(string w) {
     skipSpaces();
-    for (int i = 0; i < w.length(); i++) {
+    for (int i = 0; i < (int) w.length(); i++) {
         int j = index + i;
-        if (j >= input.length()) {
+        if (j >= (int) input.length()) {
             return false;
         }
