+++ Team41/Code41/src/spa/src/SimpleParser/Tokenizer.cpp
@@ -16,12 +16,12 @@ Tokenizer::Tokenizer(string s) : input(s), idx(0), row(0), col(0) {
 
 // private functions
 void Tokenizer::advance() {
-  if (idx >= input.size()) {
+  if (idx >= (int) input.size()) {
 	currToken = EOF_CHAR;
 	return;
   }
   advancePosition();
-  if (idx >= input.size()) {
+  if (idx >= (int) input.size()) {
 	currToken = EOF_CHAR;
   } else {
 	currToken = input[idx];
@@ -52,7 +52,7 @@ string Tokenizer::withPosition(const string &s) {
 }
 
 Token Tokenizer::eatN(TokenType type, int n) {
-  if (n > (input.size() - idx)) throw TokenizeException("eating too many characters: " + to_string(n));
+  if (n > ((int) input.size() - idx)) throw TokenizeException("eating too many characters: " + to_string(n));
 
   pair<int, int> start = {row, col};
   string val;
